snap path check

+++ src/extension/writer.ts
@@ -21,18 +21,18 @@ import { window } from "vscode";
 import { platform, release } from "os";
 import { exec } from "@vscode/sudo-prompt";
 
-import { PathOrFileDescriptor, readFileSync, writeFileSync } from "fs";
+import { PathLike, readFileSync, writeFileSync } from "fs";
 
 import { reload } from "../lib/vscode";
 import { copyCommand, generateChecksum } from "../lib/file";
 
 import { clean, inject } from "./inject"
 
-export const install: (workbench: PathOrFileDescriptor, product: PathOrFileDescriptor) => void = (workbench: PathOrFileDescriptor, product: PathOrFileDescriptor) => {
+export const install: (workbench: PathLike, product: PathLike) => void = (workbench: PathLike, product: PathLike) => {
     write(workbench, product, inject(readFileSync(workbench, "utf-8")));
 }
 
-export const uninstall: (workbench: PathOrFileDescriptor, product: PathOrFileDescriptor) => void = (workbench: PathOrFileDescriptor, product: PathOrFileDescriptor) => {
+export const uninstall: (workbench: PathLike, product: PathLike) => void = (workbench: PathLike, product: PathLike) => {
     write(workbench, product, clean(readFileSync(workbench, "utf-8")));
 }
 
@@ -40,7 +40,7 @@ export const uninstall: (workbench: PathOrFileDescriptor, product: PathOrFileDes
 
 const workbenchChecksum: RegExp = /(?<=^\s*"vs\/workbench\/workbench\.desktop\.main\.js\": \").*(?=\",\s*$)/gm;
 
-const write: (workbench: PathOrFileDescriptor, product: PathOrFileDescriptor, content: string) => void = (workbench: PathOrFileDescriptor, product: PathOrFileDescriptor, content: string) => {
+const write: (workbench: PathLike, product: PathLike, content: string) => void = (workbench: PathLike, product: PathLike, content: string) => {
     const pJson: string = readFileSync(product, "utf-8").replace(workbenchChecksum, generateChecksum(content));
 
     try{ // write changes
@@ -48,39 +48,47 @@ const write: (workbench: PathOrFileDescriptor, product: PathOrFileDescriptor, co
         writeFileSync(product, pJson, "utf-8");
         reload();
     }catch(error: any){
-        window.showWarningMessage(
-            "Failed to write changes, run command as administrator?",
-            {
-                detail: `The Background extension does not have permission to write changes, run command using administrator permissions?\n\n${error.message}`,
-                modal: true
-            },
-            "Yes"
-        ).then((value?: string) => {
-            if(value === "Yes"){
-                // use temp files
-                const workbenchTemp = fileSync().name;
-                writeFileSync(workbenchTemp, content, "utf-8");
-                const productTemp = fileSync().name;
-                writeFileSync(productTemp, pJson, "utf-8");
+        const snap: boolean = platform() === "linux" &&
+        /* also in         */ workbench.toString().replace('\\', '/').includes("/snap/") &&
+        /* extension.ts    */ product.toString().replace('\\', '/').includes("/snap/");
 
-                // sudo copy
-                const command: string = copyCommand([
-                    [workbenchTemp, workbench],
-                    [productTemp, product]
-                ]);
-                exec(command, { name: "VSCode Extension Host" }, (err: any) => {
-                    if(err)
-                        window.showErrorMessage(
-                            "Failed to write changes",
-                            {
-                                detail: `OS: ${platform()} ${release()}\nUsing command: ${command}\n\n${err.name}\n${err.message}`.trim(),
-                                modal: true
-                            }
-                        );
-                    else
-                        reload();
-                });
-            }
-        })
+        if(snap){
+            // TODO
+        }else{
+            window.showWarningMessage(
+                "Failed to write changes, run command as administrator?",
+                {
+                    detail: `The Background extension does not have permission to write changes, run command using administrator permissions?\n\n${error.message}`,
+                    modal: true
+                },
+                "Yes"
+            ).then((value?: string) => {
+                if(value === "Yes"){
+                    // use temp files
+                    const workbenchTemp = fileSync().name;
+                    writeFileSync(workbenchTemp, content, "utf-8");
+                    const productTemp = fileSync().name;
+                    writeFileSync(productTemp, pJson, "utf-8");
+
+                    // sudo copy
+                    const command: string = copyCommand([
+                        [workbenchTemp, workbench],
+                        [productTemp, product]
+                    ]);
+                    exec(command, { name: "VSCode Extension Host" }, (err: any) => {
+                        if(err)
+                            window.showErrorMessage(
+                                "Failed to write changes",
+                                {
+                                    detail: `OS: ${platform()} ${release()}\nUsing command: ${command}\n\n${err.name}\n${err.message}`.trim(),
+                                    modal: true
+                                }
+                            );
+                        else
+                            reload();
+                    });
+                }
+            });
+        }
     }
 }
